requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
requests==2.32.3
websockets==12.0
clickhouse-connect==0.7.9
python-dotenv==1.0.1
kafka-python-ng==2.2.2
sqlalchemy==2.0.49
psycopg2-binary==2.9.11
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
bcrypt==3.2.0
httpx==0.27.0
alembic==1.18.4